examples/plone/requirements-plone52-python3.txt
AccessControl==4.0
Acquisition==4.6
AuthEncoding==4.1
BTrees==4.5.1
Chameleon==3.6.2
DateTime==4.3
DocumentTemplate==3.1b2
ExtensionClass==4.4
Markdown==3.1.1
Missing==4.1
MultiMapping==4.1
Paste==3.0.8
PasteDeploy==2.0.1
Persistence==3.0
Plone==5.2.0
Products.BTreeFolder2==4.2
Products.CMFCore==2.4.0
Products.CMFDiffTool==3.3.0
Products.CMFDynamicViewFTI==6.0.2
Products.CMFEditions==3.3.2
Products.CMFFormController==4.1.0
Products.CMFPlacefulWorkflow==2.0.0
Products.CMFPlone==5.2.0
Products.CMFQuickInstallerTool==4.0.1
Products.CMFUid==3.0.1
Products.DCWorkflow==2.4.0b3
Products.DateRecurringIndex==3.0.0
Products.ExtendedPathIndex==3.4.2
Products.ExternalMethod==4.3
Products.GenericSetup==2.0
Products.MailHost==4.7
Products.MimetypesRegistry==2.1.5
Products.PlonePAS==6.0.2
Products.PluggableAuthService==2.0
Products.PluginRegistry==1.7
Products.PortalTransforms==3.1.7
Products.PythonScripts==4.7
Products.Sessions==4.5
Products.SiteErrorLog==5.3
Products.StandardCacheManagers==4.0.2
Products.TemporaryFolder==5.3
Products.ZCatalog==5.0.1
Products.ZopeVersionControl==1.1.4
Products.statusmessages==5.0.4
PyJWT==1.7.1
Record==3.5
RestrictedPython==4.0
Unidecode==0.4.1
WSGIProxy2==0.4.6
WebOb==1.8.5
WebTest==2.0.33
ZConfig==3.5.0
ZEO==5.2.1
ZODB==5.5.1
ZODB3==3.11.0
Zope==4.1.1
Zope2==4.0
beautifulsoup4==4.7.1
borg.localrole==3.1.6
calmjs.parse==1.1.3
certifi==2019.6.16
chardet==3.0.4
collective.monkeypatcher==1.2
cssselect==1.0.3
decorator==4.4.0
diazo==1.3.0
docutils==0.14
feedparser==5.2.1
five.customerize==2.0.0
five.intid==1.2.3
five.localsitemanager==3.2.2
future==0.17.1
icalendar==4.0.3
idna==2.8
jsonschema==3.0.1
mockup==3.0.2
persistent==4.5.0
piexif==1.1.2
plone.alterego==1.1.3
plone.api==1.10.0
plone.app.caching==2.0.2
plone.app.content==3.7.0
plone.app.contentlisting==2.0.0
plone.app.contentmenu==2.3.0
plone.app.contentrules==4.1.1
plone.app.contenttypes==2.1.2
plone.app.customerize==1.3.10
plone.app.dexterity==2.6.3
plone.app.discussion==3.2.1
plone.app.event==3.2.4
plone.app.i18n==3.0.5
plone.app.intid==1.1.3
plone.app.iterate==3.3.12
plone.app.layout==3.2.1
plone.app.linkintegrity==3.3.11
plone.app.locales==5.1.18
plone.app.lockingbehavior==1.0.6
plone.app.multilingual==5.3.5
plone.app.portlets==4.4.2
plone.app.querystring==1.4.11
plone.app.redirector==2.1.0
plone.app.registry==1.7.5
plone.app.relationfield==2.0.1
plone.app.textfield==1.3.1
plone.app.theming==4.0.1
plone.app.upgrade==2.0.27
plone.app.users==2.6.0
plone.app.uuid==2.0.0
plone.app.versioningbehavior==1.3.8
plone.app.viewletmanager==3.1.0
plone.app.vocabularies==4.1.0
plone.app.widgets==3.0.2
plone.app.workflow==4.0.1
plone.app.z3cform==3.1.1
plone.autoform==1.8.1
plone.batching==1.1.5
plone.behavior==1.3.0
plone.browserlayer==2.2.2
plone.cachepurging==2.0.1
plone.caching==1.2.1
plone.contentrules==2.0.9
plone.dexterity==2.9.1
plone.event==1.4.0
plone.folder==3.0.0
plone.formwidget.namedfile==2.0.9
plone.formwidget.recurrence==2.1.3
plone.i18n==4.0.5
plone.indexer==1.0.6
plone.intelligenttext==3.0.0
plone.keyring==3.1.1
plone.locking==2.2.3
plone.memoize==2.0.1
plone.namedfile==5.0.4
plone.outputfilters==3.1.2
plone.portlet.collection==3.3.1
plone.portlet.static==3.1.3
plone.portlets==2.3.1
plone.protect==4.1.2
plone.recipe.zope2instance==6.3.0
plone.registry==1.1.5
plone.resource==2.1.1
plone.resourceeditor==3.0.0
plone.rest==1.4.0
plone.restapi==4.3.1
plone.rfc822==2.0.1
plone.scale==3.0.3
plone.schema==1.2.0
plone.schemaeditor==2.0.20
plone.session==3.7.3
plone.staticresources==1.1.0
plone.stringinterp==1.3.1
plone.subrequest==1.9.1
plone.supermodel==1.4.0
plone.synchronize==1.0.3
plone.theme==3.0.6
plone.transformchain==2.0.1
plone.uuid==1.0.5
plone.z3cform==1.1.0
plonetheme.barceloneta==2.1.4
ply==3.11
pyrsistent==0.15.2
python-dateutil==2.8.0
python-gettext==4.0
pytz==2019.1
repoze.xmliter==0.6
requests==2.22.0
roman==3.2
setuptools==41.0.1
simplejson==3.16.0
soupsieve==1.9.2
tempstorage==5.0
transaction==2.4.0
urllib3==1.25.3
waitress==1.3.0
z3c.autoinclude==0.3.9
z3c.caching==2.1
z3c.form==3.7.0
z3c.formwidget.query==0.17
z3c.objpath==1.2
z3c.pt==3.2.0
z3c.relationfield==0.8.0
z3c.zcmlhook==1.0b1
zExceptions==4.1
zc.lockfile==1.4
zc.relation==1.1.post2
zdaemon==4.3
zodbpickle==1.0.4
zope.annotation==4.7.0
zope.app.locales==4.0.1
zope.browser==2.3
zope.browsermenu==4.4
zope.browserpage==4.4.0
zope.browserresource==4.3
zope.cachedescriptors==4.3.1
zope.component==4.5
zope.componentvocabulary==2.2.0
zope.configuration==4.3.1
zope.container==4.2.2
zope.contentprovider==4.2.1
zope.contenttype==4.4
zope.copy==4.2
zope.deferredimport==4.3
zope.deprecation==4.4.0
zope.dottedname==4.3
zope.event==4.4
zope.exceptions==4.3
zope.filerepresentation==4.2.0
zope.globalrequest==1.5
zope.hookable==4.2.0
zope.i18n==4.6.2
zope.i18nmessageid==4.3.1
zope.interface==4.6.0
zope.intid==4.3.0
zope.keyreference==4.2.0
zope.lifecycleevent==4.3
zope.location==4.2
zope.pagetemplate==4.4.1
zope.processlifetime==2.3.0
zope.proxy==4.3.1
zope.ptresource==4.2.0
zope.publisher==5.0.1
zope.ramcache==2.3
zope.schema==4.9.3
zope.security==4.3.1
zope.sendmail==5.0
zope.sequencesort==4.1.2
zope.site==4.2.2
zope.size==4.3
zope.structuredtext==4.3
zope.tal==4.4
zope.tales==5.0.1
zope.testbrowser==5.3.3
zope.traversing==4.3.1
zope.viewlet==4.2.1
